requirements.txt
Flask==1.1.2
Flask-Reuploaded==0.3.2
Matplotlib ==   3.5.1
Numpy ==   1.19.2
opencv-contrib-python ==   4.5.5.64
opencv-python ==   4.5.5.64
Pandas ==   1.4.2
scikit-learn ==   1.0.2
tensorflow-gpu ==   2.6.0
Imutils ==   0.5.4
PyTorch ==   1.11.0
torchvision ==   0.12.0

